+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix `RemoveIllegalComponents()` being broken when the sdk was imported via VCC.
 * Change signaling value to NaN. This fixes animated properties interpolating the signal value weirdly.
 * Create optimized swap materials after merging materials instead of before so that Texture2DArray and animated properties are correctly set.
+* Deduplicate optimized swap materials with the optimized material that is already in the mesh renderer.
 
 ## v1.9
 ### Features